+++ src/cores/storage/index.ts
@@ -1,0 +1,14 @@
+/**
+ * `cores/storage/index.tx`
+ * - default index to export all for storage-service
+ *
+ * @author      Steve Jung <steve@lemoncloud.io>
+ * @date        2022-05-13 initial file to group
+ *
+ * @copyright (C) lemoncloud.io 2022 - All Rights Reserved.
+ */
+export * from './model-manager';
+export * from './redis-storage-service';
+export * from './storage-service';
+export * from './proxy-storage-service';
+export * from './http-storage-service';
