+++ src/cores/storage/model-manager.ts
@@ -8,7 +8,7 @@
  * @copyright (C) 2020 LemonCloud Co Ltd. - All Rights Reserved.
  */
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { _log, _inf, _err, $U } from '../engine/';
+import { _log, _inf, _err, $U } from '../../engine/';
 import {
     CoreModel,
     StorageMakeable,
