+++ src/cores/storage/proxy-storage-service.ts
@@ -9,12 +9,12 @@
  * @copyright (C) 2019 LemonCloud Co Ltd. - All Rights Reserved.
  */
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { _log, _inf, _err, $U } from '../engine/';
+import { _log, _inf, _err, $U } from '../../engine/';
 import { StorageService, StorageModel } from './storage-service';
 import { DummyStorageService, DynamoStorageService } from './storage-service';
-import { NUL404 } from '../common/test-helper';
-import { Elastic6SimpleQueriable } from './core-types';
-import { GeneralAPIController } from '../controllers/general-api-controller';
+import { NUL404 } from '../../common/test-helper';
+import { Elastic6SimpleQueriable } from '../core-types';
+import { GeneralAPIController } from '../../controllers/general-api-controller';
 const NS = $U.NS('PSTR', 'blue'); // NAMESPACE TO BE PRINTED.
 
 /**
